Supports re-starting failed SAs and replaying specific SAs

+++ SASUMO/creator.py
@@ -1,6 +1,7 @@
 # from SASUMO.SASUMO.functions.function import
 import os
 import pickle
+import re
 import sys
 import random
 from importlib_metadata import distribution
@@ -28,26 +29,37 @@ SAMPLES_FILE_NAME = "SALib_Samples.txt"
 RESULTS_NAME = "output.txt"
 SOBOL_ANALYSIS = lambda x: f"sobol_analysis_{x}.csv"
 
+
 def _to_lognormal(mu, sd):
     import math
+
     lns = math.sqrt(math.log(sd / mu) ** 2 + 1)
     lnmu = math.log(mu) - 0.5 * math.log((sd / mu) ** 2 + 1)
     return lnmu, lns
 
 
 class SASUMO:
-    def __init__(self, yaml_file_path: str) -> None:
+    def __init__(self, yaml_file_path: str, rerun: bool = False, rerun_file: str = '') -> None:
         # instantate the Settings4SASUMO class
         self._settings = Config(yaml_file_path)
 
         # try to create the folder to work in.
-        create_folder(self._settings.Metadata.output)
+        try:
+            create_folder(self._settings.Metadata.output)
+            self._folder_exists = False
+        except FileExistsError:
+            self._folder_exists = True
+            if rerun:
+                self._rerun = True
+                with open(rerun_file, "r") as f:
+                    self._rerun_list = f.read().splitlines(keepends=False)
 
         # save a copy of the settings file
-        self._settings.to_yaml(
-            os.path.join(self._settings.Metadata.output, "sasumo_params.yaml"),
-            resolve=True,
-        )
+        if not self._folder_exists:
+            self._settings.to_yaml(
+                os.path.join(self._settings.Metadata.output, "sasumo_params.yaml"),
+                resolve=True,
+            )
 
         # generate and save the problem definition
         self._problem = self._compose_problem()
@@ -56,6 +68,19 @@ class SASUMO:
 
         # generate the samples
         self._samples = self._generate_samples()
+        if self._folder_exists:
+            if not self._rerun:
+                sp = self._find_start_point()
+                self._sp = lambda x: sp + x
+                print(f"[!] Starting from incomplete SA. Continuing from {self._sp(0)}")
+                self._samples = self._samples[self._sp:]
+            else:
+                print(f"[!] Starting from incomplete SA. Continuing from {self._rerun_list[0]}")
+                self._remove_incomplete(self._rerun_list)
+                self._samples = [self._samples[int(s)] for s in self._rerun_list]
+                self._sp = lambda x: int(self._rerun_list[x])  # the start point is the last rerun
+        else:
+            self._sp = lambda x: 0 + x  # the start point is 0
 
         # add the required paths to PYTHON PATH
         self._update_path()
@@ -84,6 +109,31 @@ class SASUMO:
             if new_path:
                 sys.path.append(new_path)
 
+    def _remove_incomplete(self, incomplete_list) -> None:
+        import shutil
+        for dir in incomplete_list:
+            shutil.rmtree(os.path.join(self._settings.Metadata.output, str(dir)))
+
+
+    def _find_start_point(self) -> int:
+        import shutil
+
+        incomplete = [
+            int(_dir.name)
+            for _dir in os.scandir(self._settings.Metadata.output)
+            # TODO: f_out is a hard code and will not work for others
+            if _dir.is_dir() and "f_out.txt" not in os.listdir(_dir.path)
+        ]
+        incomplete.sort()
+        if incomplete:
+            sp = int(incomplete[0])
+            if len(incomplete) > 1:
+                for dir in range(sp, int(incomplete[-1]) + 1):
+                    shutil.rmtree(os.path.join(self._settings.Metadata.output, str(dir)))
+        else:
+            sp = sorted((int(_dir.name) for _dir in os.scandir(self._settings.Metadata.output) if _dir.is_dir()), reverse=True)[0] + 1
+        return sp
+
     def _generate_seed(
         self,
     ):
@@ -95,6 +145,11 @@ class SASUMO:
     def _generate_samples(
         self,
     ) -> np.array:
+        if self._folder_exists:
+            # load the samples from the file
+            return np.loadtxt(
+                os.path.join(self._settings.Metadata.output, SAMPLES_FILE_NAME)
+            )
 
         if self._settings.SensitivityAnalysis.get("mode", "") == "correlated":
             from SALib.sample import sobol_corr
@@ -123,42 +178,52 @@ class SASUMO:
     def _compose_problem(
         self,
     ) -> dict:
-
-        return {
-            "num_vars": len(self._settings.SensitivityAnalysis.Variables),
-            "names": [
-                name
-                for name, var in self._settings.SensitivityAnalysis.Variables.items()
-            ],
-            "bounds": [
-                self._compose_bounds(var)
-                for var in self._settings.SensitivityAnalysis.Variables.values()
-            ],
-            "dists": [
-                var.get("sobol_dist", "unif")
-                for var in self._settings.SensitivityAnalysis.Variables.values()
-            ],
-            **(
-                {
-                    "distrs": [
-                        var.distr
-                        for _, var in self._settings.SensitivityAnalysis.Variables.items()
-                    ]
-                }
-                if self._settings.SensitivityAnalysis.get("mode", "") == "correlated"
-                else {}
-            ),
-            **(
-                {
-                    "corr": [
-                        var.corr
-                        for _, var in self._settings.SensitivityAnalysis.Variables.items()
-                    ]
-                }
-                if self._settings.SensitivityAnalysis.get("mode", "") == "correlated"
-                else {}
-            ),
-        }
+        if self._folder_exists:
+            with open(
+                os.path.join(
+                    self._settings.Metadata.output, PROBLEM_DESCRIPT_FILE_NAME
+                ),
+                "r",
+            ) as f:
+                return json.load(f)
+        else:
+            return {
+                "num_vars": len(self._settings.SensitivityAnalysis.Variables),
+                "names": [
+                    name
+                    for name, var in self._settings.SensitivityAnalysis.Variables.items()
+                ],
+                "bounds": [
+                    self._compose_bounds(var)
+                    for var in self._settings.SensitivityAnalysis.Variables.values()
+                ],
+                "dists": [
+                    var.get("sobol_dist", "unif")
+                    for var in self._settings.SensitivityAnalysis.Variables.values()
+                ],
+                **(
+                    {
+                        "distrs": [
+                            var.distr
+                            for _, var in self._settings.SensitivityAnalysis.Variables.items()
+                        ]
+                    }
+                    if self._settings.SensitivityAnalysis.get("mode", "")
+                    == "correlated"
+                    else {}
+                ),
+                **(
+                    {
+                        "corr": [
+                            var.corr
+                            for _, var in self._settings.SensitivityAnalysis.Variables.items()
+                        ]
+                    }
+                    if self._settings.SensitivityAnalysis.get("mode", "")
+                    == "correlated"
+                    else {}
+                ),
+            }
 
     def _compose_bounds(self, variable_obj: object) -> Tuple[float, float]:
         """
@@ -174,57 +239,66 @@ class SASUMO:
             # return (variable_obj.distribution.get("min", 0), variable_obj.distribution.get("max", 1))
             try:
                 return (
-                variable_obj.get("sobol_dist_params", ).get("lb", 0),
-                variable_obj.get("sobol_dist_params", ).get("ub"),
-            )
+                    variable_obj.get(
+                        "sobol_dist_params",
+                    ).get("lb", 0),
+                    variable_obj.get(
+                        "sobol_dist_params",
+                    ).get("ub"),
+                )
             except (TypeError, AttributeError):
                 return (
                     variable_obj.sumo_dist.params.get("lb", 0),
                     variable_obj.sumo_dist.params.ub,
                 )
-        elif variable_obj.get("sobol_dist", "unif") in ["norm", ]:
+        elif variable_obj.get("sobol_dist", "unif") in [
+            "norm",
+        ]:
             return (
                 variable_obj.sobol_dist_params.mean,
                 variable_obj.sobol_dist_params.std,
             )
         elif variable_obj.get("sobol_dist", "unif") == "lognorm":
-            return _to_lognormal(variable_obj.sobol_dist_params.mean, variable_obj.sobol_dist_params.std)
-            
+            return _to_lognormal(
+                variable_obj.sobol_dist_params.mean, variable_obj.sobol_dist_params.std
+            )
+
         raise NotImplementedError("This distribution is not supported")
 
     def _save_problem(
         self,
     ) -> None:
-
-        with open(
-            os.path.join(
-                self._settings.Metadata.output,
-                PROBLEM_DESCRIPT_FILE_NAME,
-            ),
-            "w",
-        ) as f:
-            f.write(
-                json.dumps(
-                    self._problem,
-                    indent=4,
-                    quote_keys=True,
-                    trailing_commas=False,
+        if not self._folder_exists:
+            with open(
+                os.path.join(
+                    self._settings.Metadata.output,
+                    PROBLEM_DESCRIPT_FILE_NAME,
+                ),
+                "w",
+            ) as f:
+                f.write(
+                    json.dumps(
+                        self._problem,
+                        indent=4,
+                        quote_keys=True,
+                        trailing_commas=False,
+                    )
                 )
-            )
 
     def main(self, smoke_test=False) -> List[List[float]]:
 
         dispatch = []
         results = []
 
-        # dispatch = [(i, self._spawn_process(i)) for i in range(self._settings.sensitivity_analysis.num_runs)]
+        num_parallel = min(self._settings.SensitivityAnalysis.parallel_trials, os.cpu_count(), len(self._samples))
+
         for i, _ in enumerate((range(2),) * 2 if smoke_test else self._samples):
 
             dispatch.append([i, self._spawn_process(i)])
 
             while len(
                 dispatch
-            ) >= self._settings.SensitivityAnalysis.parallel_trials or (
+            ) >= num_parallel or (
                 i >= (len(self._samples) - 1) and dispatch
             ):
 
@@ -247,7 +321,7 @@ class SASUMO:
         p = self._f.remote(
             self._settings.generate_process(
                 process_var=self._samples[index],
-                process_id=str(index),
+                process_id=str(self._sp(index)),
                 random_seed=self._generate_seed(),
             )
         )
@@ -263,37 +337,41 @@ class SASUMO:
         self._f(
             yaml_params=self._settings.generate_process(
                 process_var=self._samples[index],
-                process_id=str(index),
+                process_id=str(self._sp(index)),
                 random_seed=self._generate_seed(),
             ),
         ).run()
 
     def save_results(self, sobol_analysis: list, results: list) -> None:
         # save the sobol analysis
-        for i, result in enumerate(sobol_analysis.to_df()):
-            result.to_csv(
-                os.path.join(self._settings.Metadata.output, SOBOL_ANALYSIS(i))
-            )
 
-        # save the results
-        np.savetxt(
-            os.path.join(self._settings.Metadata.output, RESULTS_NAME),
-            np.array(results),
-        )
+        if not self._folder_exists:
+        
+            for i, result in enumerate(sobol_analysis.to_df()):
+                result.to_csv(
+                    os.path.join(self._settings.Metadata.output, SOBOL_ANALYSIS(i))
+                )
+
+            # save the results
+            np.savetxt(
+                os.path.join(self._settings.Metadata.output, RESULTS_NAME),
+                np.array(results),
+            )
 
     def analyze(self, results) -> dict:
-        if self._settings.SensitivityAnalysis.get("mode", "") == "correlated":
-            from SALib.analyze import sobol_corr
+        if not self._folder_exists:
+            if self._settings.SensitivityAnalysis.get("mode", "") == "correlated":
+                from SALib.analyze import sobol_corr
 
-            return sobol_corr.analyze(
-                self._problem,
-                np.array([r[0] for r in results]),
-                self._settings.SensitivityAnalysis.N
+                return sobol_corr.analyze(
+                    self._problem,
+                    np.array([r[0] for r in results]),
+                    self._settings.SensitivityAnalysis.N,
+                )
+
+            return sobol.analyze(
+                self._problem, np.array([r[0] for r in results]), print_to_console=True
             )
-        
-        return sobol.analyze(
-            self._problem, np.array([r[0] for r in results]), print_to_console=True
-        )
 
 
 @click.command()
@@ -308,10 +386,16 @@ class SASUMO:
     is_flag=True,
     help="Finish a sensitivity analysis that quit for some reason. It will replay the last # - CPU * 2, just to be safe",
 )
+@click.option(
+    "--rerun", is_flag=True, help="Re-run the sensitivity analysis for specified iterations"
+)
+@click.option(
+    "--rerun-file", type=click.Path(exists=True), help="file of iterations to re-run, deliminated by newlines"
+)
 @click.argument("settings_file")
-def run(debug, smoke_test, finish_existing, settings_file):
+def run(debug, smoke_test, finish_existing, settings_file, rerun, rerun_file):
 
-    s = SASUMO(settings_file)
+    s = SASUMO(settings_file, rerun, rerun_file)
 
     if debug:
         if "Remote" in s._settings.get("ManagerFunction").module:
